% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bp}
\alias{bp}
\title{MCMC with bipartite graph}
\usage{
bp(whole, part, edge, alpha = 0.05, gamma = 0.2, p = 0.01,
  nburn = 10000, ngen = 100000, sub = 1000, penalty = 2,
  initial = c("inactive", "random", "high"))
}
\arguments{
\item{whole}{Vector of character strings with names of whole nodes (gene-sets)}

\item{part}{Vector of 0's and 1's indicating active part nodes; names
  are the names of the part nodes (genes)}

\item{edge}{Matrix with two columns; each row indicates an edge between a whole node and a part node}

\item{alpha}{Parameter alpha (0 < alpha < gamma < 1)}

\item{gamma}{Parameter gamma (0 < alpha < gamma < 1)}

\item{p}{Parameter p (0 < p < 1)}

\item{nburn}{Number of burn-in generations}

\item{ngen}{Number of sample generations}

\item{sub}{Subsample rate for burn-in and sample files}

\item{penalty}{Penalty per illegal node to loglikelihood}

\item{initial}{Initial state (see Details)}
}
\value{
data frame with "whole" results; 
\item{Name}{The names of whole nodes}
\item{ActiveProbability}{The marginal posterior probabilities of whole nodes}
\item{Count}{Number of active samples for each whole node}
\item{Sample}{Number of generated samples for each whole node}
\item{Degree}{The degree of each whole node}
\item{Response}{The response of each whole node}
dimburn + sample detail are included as an attribute, \code{"samples"}, which is a matrix.
}
\description{
MCMC with bipartite graph
}
\details{
The \code{initial} argument can take one of three values:
\code{"inactive"} - all whole nodes inactive; \code{"random"} - all
whole nodes active with probability \code{p}, no illegal nodes; or
\code{"high"} - all nodes with proportion of connected part nodes
with response equal to 1 above 0.4 are active, no illegal nodes.
}
\examples{
data(t2d)

## eout <- I2edge(t2d$I)

## bp.out <- bp(whole=colnames(t2d$I), part=t2d$y, edge=eout,
## alpha=0.00019, gamma=0.023, p=0.0033,
## nburn=1000000, ngen=10000000, sub=1000, penalty=5, initial="random")
}
\keyword{models}

