% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bp}
\alias{bp}
\title{MCMC with bipartite graph}
\usage{
bp(whole, part, edge, alpha = 0.05, beta = 0.2, pi = 0.01,
  nburn = 10000, ngen = 100000, sub = 1000, penalty = 2,
  initial = c("inactive", "random", "high"))
}
\arguments{
\item{whole}{Vector of character strings with names of whole nodes}

\item{part}{Vector of 0's and 1's indicating active part nodes; names are the names of the part nodes}

\item{edge}{Matrix with two columns; each row indicates an edge between a whole node and a part node}

\item{alpha}{Parameter alpha (0 < alpha < beta < 1)}

\item{beta}{Parameter alpha (0 < alpha < beta < 1)}

\item{pi}{Parameter pi (0 < pi < 1)}

\item{nburn}{Number of burn-in generations}

\item{ngen}{Number of sample generations}

\item{sub}{Subsample rate for burn-in and sample files}

\item{penalty}{Penalty per illegal node to loglikelihood}

\item{initial}{Initial state (see Details)}
}
\value{
data frame with "whole" results; burn+sample detail are included as an attribute,
\code{"samples"}, which is a matrix
}
\description{
MCMC with bipartite graph
}
\details{
The \code{initial} argument can take one of three values:
\code{"inactive"} - all whole nodes inactive; \code{"random"} - all
whole nodes active with probability \code{pi}, no illegal nodes; or
\code{"high"} - all nodes with proportion of connected part nodes
with response equal to 1 above 0.4 are active, no illegal nodes.
}
\examples{
data(t2d)
bp.out <- bp(whole=t2d$whole, part=t2d$part, edge=t2d$edge,
             nburn=1000, ngen=1000, sub=100)
}
\keyword{models}

